support/doc-templates/assets/archive/eodeck-starter/eodeck-starter/requirements.txt
python-pptx==0.6.22
Pillow>=10.0.0
PyYAML>=6.0.0
